AddBitsLen() use the uint32/64_t sha-512 need 16 bytes(uint64_t[2]) for msg length padding

+++ Hash/sha/sha256.c
@@ -72,7 +72,7 @@ ErrCrypto SHA256_init(HashState* pHashState)
 
 
 
-ErrCrypto AddBitsLen(HashState* pHashState, uint16_t nBits)
+ErrCrypto AddBitsLen(HashState* pHashState, uint32_t nBits)
 {
     // Maximum message length is 2**64 bits 
     pHashState->nBitsLen += nBits;
